+++ src/Router.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
 
-import EmailCertifyPage from "pages/auth_pages/email_certify_page/EmailCertifyPage";
-import SignInPage from "pages/auth_pages/sign_in_page/SignInPage";
+import SignInPage from "pages/auth_pages/sign_in_pages/SignInPage";
+import EmailCertifyPage from "pages/auth_pages/sign_in_pages/email_certify_page/EmailCertifyPage";
+import SignInInfoPage from "pages/auth_pages/sign_in_pages/sign_in_page/SignInInfoPage";
 import SignUpPage from "pages/auth_pages/sign_up_page/SignUpPage";
 import MainPage from "pages/main_page/MainPage";
 import MyPage from "pages/my_page/MyPage";
@@ -11,9 +12,11 @@ const Router = () => {
   return (
     <Routes>
       <Route path="/main" element={<MainPage />} />
-      <Route path="/signin" element={<SignInPage />} />
       <Route path="/signup" element={<SignUpPage />} />
-      <Route path="/signin/email" element={<EmailCertifyPage />} />
+      <Route path="/signin" element={<SignInPage />}>
+        <Route path="info" element={<SignInInfoPage />} />
+        <Route path="auth" element={<EmailCertifyPage />} />
+      </Route>
       <Route path="/mypage" element={<MyPage />} />
     </Routes>
   );
