+++ src/assets/images/icons/KeyIcon.tsx
@@ -1,0 +1,32 @@
+import React from "react";
+
+const KeyIcon = ({
+  width = 24,
+  height = 24,
+  color = "var(--color_brand01)",
+}) => {
+  return (
+    <svg
+      width={width}
+      height={height}
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M15.68 5.34799C14.8978 5.3472 14.1473 5.6571 13.5935 6.20955C13.0397 6.76201 12.728 7.51178 12.727 8.29399C12.7283 9.07621 13.0403 9.82588 13.5942 10.3781C14.1482 10.9304 14.8988 11.2401 15.681 11.239C16.4632 11.2401 17.2138 10.9304 17.7678 10.3781C18.3217 9.82588 18.6337 9.07621 18.635 8.29399C18.6345 7.90659 18.5576 7.52309 18.4089 7.16538C18.2602 6.80767 18.0424 6.48276 17.7681 6.2092C17.4938 5.93564 17.1683 5.71878 16.8102 5.57102C16.4521 5.42325 16.0674 5.34747 15.68 5.34799ZM14.228 8.29399C14.2285 8.10358 14.2665 7.91513 14.3399 7.73941C14.4132 7.56369 14.5205 7.40413 14.6555 7.26986C14.7905 7.13559 14.9507 7.02922 15.1268 6.95684C15.3029 6.88445 15.4916 6.84747 15.682 6.84799C16.488 6.84799 17.136 7.49799 17.136 8.29399C17.1347 8.67839 16.9808 9.04652 16.7081 9.31748C16.4355 9.58845 16.0664 9.74006 15.682 9.739C15.2976 9.74006 14.9285 9.58845 14.6559 9.31748C14.3832 9.04652 14.2293 8.67839 14.228 8.29399Z"
+        fill={color}
+      />
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M9.52999 20.878C9.82004 20.4866 9.97198 20.01 9.96199 19.523C10.354 19.639 10.742 19.653 11.114 19.583C11.882 19.437 12.451 18.951 12.753 18.551L12.758 18.544L12.763 18.537C12.9734 18.2501 13.118 17.9204 13.1865 17.5712C13.2549 17.2221 13.2455 16.8621 13.159 16.517C13.0851 16.2163 12.9636 15.9293 12.799 15.667C12.7229 15.5265 12.6317 15.3947 12.527 15.274L12.536 15.264L13 14.802C13.382 14.985 13.796 15.1 14.158 15.175C14.6593 15.2776 15.1693 15.3318 15.681 15.337C19.582 15.337 22.75 12.187 22.75 8.293C22.75 4.401 19.582 1.25 15.68 1.25C11.78 1.25 8.61099 4.4 8.61099 8.293C8.61476 9.20152 8.80893 10.0992 9.18099 10.928L1.92499 18.154C1.58646 18.4968 1.36042 18.9345 1.27699 19.409C1.17799 20.012 1.33999 20.683 1.92499 21.265L2.80699 22.143L2.84899 22.183C3.19125 22.4696 3.60503 22.6576 4.04599 22.727C4.36006 22.7756 4.68122 22.748 4.98233 22.6463C5.28345 22.5447 5.55567 22.372 5.77599 22.143L6.34199 21.579C6.82972 21.864 7.40436 21.962 7.95899 21.855C8.578 21.7323 9.13193 21.3904 9.51899 20.892L9.52499 20.885L9.52999 20.878ZM10.111 8.293C10.111 5.235 12.602 2.75 15.681 2.75C18.759 2.75 21.25 5.235 21.25 8.293C21.25 11.353 18.76 13.837 15.68 13.837C15.408 13.837 14.937 13.804 14.462 13.707C13.965 13.603 13.575 13.45 13.367 13.277C13.2232 13.158 13.0403 13.0969 12.8539 13.1056C12.6675 13.1143 12.4911 13.1922 12.359 13.324L11.477 14.202C11.276 14.402 11.082 14.63 10.979 14.893C10.849 15.226 10.891 15.546 11.039 15.813C11.127 15.97 11.269 16.12 11.328 16.183L11.348 16.204C11.418 16.281 11.455 16.331 11.483 16.387L11.51 16.44L11.54 16.482L11.547 16.492C11.5563 16.5073 11.5703 16.5327 11.589 16.568C11.627 16.641 11.671 16.744 11.699 16.865C11.7354 16.9994 11.7409 17.1404 11.7149 17.2773C11.6889 17.4141 11.6322 17.5433 11.549 17.655C11.409 17.838 11.142 18.051 10.835 18.109C10.577 18.158 10.157 18.126 9.59699 17.569C9.45641 17.4289 9.266 17.3502 9.06749 17.3502C8.86899 17.3502 8.67858 17.4289 8.53799 17.569L8.24399 17.861C8.10808 17.9961 8.02904 18.1781 8.02309 18.3696C8.01714 18.5612 8.08473 18.7477 8.21199 18.891C8.22399 18.907 8.24233 18.9337 8.26699 18.971C8.31499 19.046 8.37099 19.151 8.41099 19.271C8.48699 19.496 8.49899 19.737 8.31599 19.997C8.19599 20.152 7.95299 20.329 7.67699 20.381C7.44299 20.426 7.14299 20.393 6.80499 20.056C6.66441 19.9159 6.474 19.8372 6.27549 19.8372C6.07699 19.8372 5.88658 19.9159 5.74599 20.056L4.71699 21.081C4.55199 21.245 4.40699 21.264 4.26499 21.243C4.11097 21.2147 3.96574 21.1507 3.84099 21.056L2.98399 20.202C2.74499 19.965 2.73599 19.782 2.75699 19.652C2.78918 19.4879 2.86799 19.3365 2.98399 19.216L10.628 11.603C10.7478 11.4836 10.8234 11.327 10.8425 11.1589C10.8616 10.9908 10.823 10.8212 10.733 10.678L10.731 10.675C10.7015 10.6262 10.6742 10.5761 10.649 10.525C10.299 9.8328 10.1148 9.06866 10.111 8.293ZM8.20099 18.876C8.20099 18.8773 8.20233 18.879 8.20499 18.881L8.20399 18.88L8.20099 18.876Z"
+        fill={color}
+      />
+    </svg>
+  );
+};
+
+export default KeyIcon;
