+++ src/components/inputs/base_input/BaseInput.module.scss
@@ -9,4 +9,8 @@
   &::placeholder {
     color: $grey04;
   }
+
+  &:focus {
+    outline: 1px solid $brand01;
+  }
 }
