--- src/pages/auth_pages/email_certify_page/EmailCertifyPage.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const EmailCertifyPage = () => {
-  return <div></div>;
-};
-
-export default EmailCertifyPage;
