+++ src/pages/auth_pages/sign_in_pages/email_certify_page/EmailCertifyPage.module.scss
@@ -1,0 +1,37 @@
+@import "styles/variables";
+
+.page_wrap {
+  @include flex(column, center, center, 40px);
+  width: 100%;
+  height: calc(100vh - 60px);
+}
+
+.email_auth_section {
+  @include flex(column, start, center, 20px);
+  width: 100%;
+  padding: 0 20px;
+}
+
+.email_auth_title_wrap {
+  @include flex(row, center, center, 10px);
+}
+
+.email_auth_wrap {
+  @include flex(row, center, center, 10px);
+}
+
+.email_auth_title {
+  color: $brand01;
+}
+
+.email_auth_desc {
+  text-align: center;
+}
+
+.email_notice_wrap {
+  @include flex(row, center, center, 8px);
+
+  > span {
+    font-size: $fs;
+  }
+}
