+++ src/pages/main_page/MainPage.tsx
@@ -8,7 +8,7 @@ const MainPage = () => {
   const navigate = useNavigate();
 
   const handleSignInClick = () => {
-    navigate("/signin");
+    navigate("/signin/info");
   };
 
   const handleSignUpClick = () => {
@@ -20,15 +20,9 @@ const MainPage = () => {
       {/* 첫번째 소개 Section */}
       <section className={classes.first_section}>
         <h1 className={classes.font_center}>
-          기술 면접 질문을 매일매일 <br />
-          메일로 보내드릴게요!
+          공식문서 기반 기술 질문을 <br />
+          메일로 받아보세요!
         </h1>
-        <h3 className={`${classes.font_label} ${classes.font_center}`}>
-          따로 시간 내지 않아도, <br />
-          지하철에서 하나씩 읽다보면 면접
-          <br />
-          걱정이 사라질 거예요.
-        </h3>
         <TextButton
           text="로그인"
           size="h2"
