+++ src/styles/common.scss
@@ -18,4 +18,8 @@
     font-size: $fh3;
     font-weight: 500;
   }
+
+  strong {
+    color: $brand01;
+  }
 }
